requirements.txt
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.10.0
pynput>=1.7.6
pyperclip>=1.8.2
pystray>=0.19.4
Pillow>=10.0.0
customtkinter>=5.2.0
